Move test util for headless tests
